[IK-159] refactor:  통합테스트 네이밍 변경
